kill the unresolved issues section so nobody has to maintain it
Unlike the Kcom issue where UK users came into the discord to inform us the issue was resolved, no such thing happened for users in the Philippians, on top of while this issue originally arose on the 3rd of December, a user asked about it on the 27th.

Since the resolution of issues for users in The Philippines is unknown, I've gone ahead and intentionally worded it differently than all other previously known issues, hopefully allowing users who may still have this issue, to see it on the list and follow the available options listed.